fix(callView): move logic from Vuex to Pinia
