+++ src/store/storeConfig.js
@@ -5,7 +5,6 @@
 
 import actorStore from './actorStore.js'
 import audioRecorderStore from './audioRecorderStore.js'
-import callViewStore from './callViewStore.js'
 import conversationsStore from './conversationsStore.js'
 import fileUploadStore from './fileUploadStore.js'
 import messagesStore from './messagesStore.js'
@@ -16,7 +15,6 @@ export default {
 	modules: {
 		actorStore,
 		audioRecorderStore,
-		callViewStore,
 		conversationsStore,
 		fileUploadStore,
 		messagesStore,
